+++ languagetool-standalone/CHANGES.md
@@ -6,6 +6,8 @@
   * added and improved rules
   * New rule that checks coherent use of Du/du, Dich/dich etc. Assumes that the first
     use has 'correct' capitalization and suggest the same capitalization for subsequent uses.
+  * New line extension '-*' for 'ignore.txt': entries ending with '*-' are ignored only if 
+    they are part of a hyphenated compound
 
 
 ## 4.0 (2017-12-29)